# swagger.yaml
openapi: 3.0.0
info:
  title: Example Garage Management API
  version: "1.0.0"
  description: >
    This is a sample Swagger/OpenAPI documentation for the combined routes you posted.
servers:
  - url: http://localhost:5000
    description: Local dev server
paths:
...
  /tasks/user:
    get:
      summary: Get tasks assigned to current user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    description:
                      type: string
                    dueDate:
                      type: string
                      format: date-time
                    status:
                      type: string
                    vehicle:
                      type: object
                      properties:
                        brand:
                          type: string
                        model:
                          type: string
        "500":
          description: Server error
  /tasks/garage/{garageId}/vehicle/{vehicleId}:
    post:
      summary: Create a new task for a vehicle in a specific garage (admin only)
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: garageId
          required: true
          schema:
            type: string
        - in: path
          name: vehicleId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description, dueDate]
              properties:
                description:
                  type: string
                dueDate:
                  type: string
                  format: date-time
                assignedTo:
                  type: string
      responses:
        "201":
          description: Task created
        "403":
          description: Access denied (must be the admin of that garage)
        "500":
          description: Server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []